+++ core/helpers/initCacheStorage.ts
@@ -1,7 +1,7 @@
 import * as localForage from 'localforage'
-import UniversalStorage from '@vue-storefront/store/lib/storage'
+import UniversalStorage from '@vue-storefront/core/store/lib/storage'
 import { currentStoreView } from '@vue-storefront/core/lib/multistore'
-import rootStore from '@vue-storefront/store'
+import rootStore from '@vue-storefront/core/store'
 
 /** Inits cache storage for given module. By default via local storage */
 export function initCacheStorage(key, localised = true) {