+++ core/modules/checkout/store/payment/index.ts
@@ -1,7 +1,7 @@
 import { Module } from 'vuex'
 import RootState from '@vue-storefront/core/types/RootState'
 import PaymentState from '../../types/PaymentState'
-import rootStore from '@vue-storefront/store'
+import rootStore from '@vue-storefront/core/store'
 
 export const paymentModule: Module<PaymentState, RootState> = {
   namespaced: true,