+++ src/themes/default/components/core/ProductTile.vue
@@ -57,7 +57,7 @@
 </template>
 
 <script>
-import rootStore from '@vue-storefront/store'
+import rootStore from '@vue-storefront/core/store'
 import { ProductTile } from '@vue-storefront/core/modules/catalog/components/ProductTile.ts'
 
 export default {